Use util functions that got moved to the koperator/api module
