+++ config/samples/kraft/simplekafkacluster_kraft.yaml
@@ -93,7 +93,7 @@ spec:
 #      limits:
 #        cpu: 500m
 #        memory: 1Gi
-#    image: "ghcr.io/banzaicloud/cruise-control:2.5.86"
+#    image: "ghcr.io/banzaicloud/cruise-control:2.5.123"
     config: |
       # Copyright 2017 LinkedIn Corp. Licensed under the BSD 2-Clause License (the "License"). See License in the project root for license information.
       #
